+++ record_service/alembic/versions/497b94b576a5_add_creation_tables.py
@@ -1,0 +1,41 @@
+"""Add creation tables
+
+Revision ID: 497b94b576a5
+Revises: 
+Create Date: 2018-06-19 22:21:15.155202
+
+"""
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision = '497b94b576a5'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('records',
+                    sa.Column('id', postgresql.UUID(
+                        as_uuid=True), nullable=False),
+                    sa.Column('archived', sa.Boolean(), nullable=False),
+                    sa.Column('record_hash', sa.Text(), nullable=False),
+                    sa.Column('created', sa.DateTime(), nullable=False),
+                    sa.PrimaryKeyConstraint('id')
+                    )
+    op.create_table('users',
+                    sa.Column('id', sa.Text(), nullable=False),
+                    sa.Column('hashed_password', sa.Text(), nullable=False),
+                    sa.PrimaryKeyConstraint('id')
+                    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('users')
+    op.drop_table('records')
+    # ### end Alembic commands ###
