--- record_service/alembic/versions/dcc34dfee532_add_creation_tables.py
@@ -1,44 +1,0 @@
-"""Add creation tables
-
-Revision ID: dcc34dfee532
-Revises: 
-Create Date: 2018-06-19 16:00:07.202452
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = 'dcc34dfee532'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('users',
-    sa.Column('id', sa.Text(), nullable=False),
-    sa.Column('password_hash', sa.Text(), nullable=False),
-    sa.Column('password_salt', sa.Text(), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('records',
-    sa.Column('id', postgresql.UUID(as_uuid=True), nullable=False),
-    sa.Column('user_id', sa.Text(), nullable=False),
-    sa.Column('archived', sa.Boolean(), nullable=False),
-    sa.Column('record_hash', sa.Text(), nullable=False),
-    sa.Column('created', sa.DateTime(), nullable=False),
-    sa.Column('acl_id', sa.Text(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('records')
-    op.drop_table('users')
-    # ### end Alembic commands ###
