+++ record_service/record_service/models/record.py
@@ -8,8 +8,6 @@ class Record(Base):
     __tablename__ = "records"
 
     id = db.Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
-    user_id = db.Column(db.Text, db.ForeignKey("users.id"), nullable=False)
     archived = db.Column(db.Boolean, nullable=False, default=False)
     record_hash = db.Column(db.Text, nullable=False)
     created = db.Column(db.DateTime, nullable=False)
-    acl_id = db.Column(db.Text, nullable=False)
